day92-image_color_palette_generator/requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.1
numpy~=2.1.0
pillow~=10.4.0